`CommonUi` -> `BrowserWrapper` 이동
- 두 파일의 경계가 모호해 하나로 합침
