+++ demo/src/components/links/ExternalLink.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { redirect } from '../../../../src/util/CommonUi';
+import { redirect } from '../../../../src/util/BrowserWrapper';
 
 
 const ExternalLink = (props) => {
