+++ src/components/screen/ImageScrollScreen.jsx
@@ -22,7 +22,7 @@ import { FOOTER_INDEX } from '../../constants/CalculationsConstants';
 import DOMEventConstants from '../../constants/DOMEventConstants';
 import DOMEventDelayConstants from '../../constants/DOMEventDelayConstants';
 import { READERJS_CONTENT_WRAPPER } from '../../index';
-import { addEventListener, removeEventListener } from '../../util/CommonUi';
+import { addEventListener, removeEventListener } from '../../util/BrowserWrapper';
 
 class ImageScrollScreen extends BaseScreen {
   componentDidMount() {
